Reorganize content classes packages
